Tao cac lop Repo, Service, Impl, Controller của các Entity

+++ src/main/java/com/example/nedcinema/Impl/UserServiceImpl.java
@@ -1,9 +1,9 @@
 package com.example.nedcinema.Impl;
 
 import com.example.nedcinema.entity.User;
+import com.example.nedcinema.Exception.*;
 import com.example.nedcinema.repository.UserRepository;
 import com.example.nedcinema.service.UserService;
-import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,13 +14,15 @@ import java.util.Optional;
 @Service
 @Transactional
 public class UserServiceImpl implements UserService {
-
     @Autowired
     private UserRepository userRepository;
 
     @Override
     public User createUser(User user) {
         // Check if the email already exists in the database
+        if(user.getRoles() == null) {
+            throw new InternalServerErrorException("Role must not be null");
+        }
 
         // Create the user
         return userRepository.save(user);
@@ -32,18 +34,27 @@ public class UserServiceImpl implements UserService {
         if (optionalUser.isPresent()) {
             User existingUser = optionalUser.get();
             existingUser.setEmail(user.getEmail());
-
             existingUser.setPassword(user.getPassword());
             existingUser.setDisplayName(user.getDisplayName());
+
+            if(user.getRoles() != null) {
+                existingUser.setRoles(user.getRoles());
+            }
+
             return userRepository.save(existingUser);
         } else {
-            throw new RuntimeException("User not found with id: " + userUID);
+            throw new NotFoundException("User not found with id: " + userUID);
         }
     }
 
     @Override
     public void deleteUser(String userUID) {
-        userRepository.deleteById(userUID);
+        Optional<User> optionalUser = userRepository.findById(userUID);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(userUID);
+        } else {
+            throw new NotFoundException("User not found with id: " + userUID);
+        }
     }
 
     @Override
@@ -52,7 +63,7 @@ public class UserServiceImpl implements UserService {
         if (optionalUser.isPresent()) {
             return optionalUser.get();
         } else {
-            throw new RuntimeException("User not found with id: " + userUID);
+            throw new NotFoundException("User not found with id: " + userUID);
         }
     }
 
@@ -60,4 +71,5 @@ public class UserServiceImpl implements UserService {
     public List<User> getAllUsers() {
         return userRepository.findAll();
     }
+
 }
